+++ CHANGELOG.md
@@ -5,4 +5,5 @@ There is a changelog available for each package:
 - [secsync](./packages/secsync/CHANGELOG.md)
 - [secsync-react-yjs](./packages/secsync-react-yjs/CHANGELOG.md)
 - [secsync-react-automerge](./packages/secsync-react-automerge/CHANGELOG.md)
+- [secsync-react-devtool](./packages/secsync-react-devtool/CHANGELOG.md)
 - [tiptap-extension-y-awareness](./packages/tiptap-extension-y-awareness/CHANGELOG.md)